+++ src/components/Header/Header.test.js
@@ -2,10 +2,8 @@ import React from "react";
 import { fireEvent, render, screen, waitFor } from "@testing-library/react"
 import { Provider } from 'react-redux';
 import configureStore from 'redux-mock-store';
-import axios from "axios";
-import { showSavedMessage } from "../../utils/Notifications";
-
 import Header from "./Header";
+import { saveProject } from "../Editor/EditorSlice";
 
 jest.mock('axios');
 
@@ -16,33 +14,35 @@ jest.mock('react-router-dom', () => ({
   })
 }));
 
-jest.mock('../../utils/Notifications')
+jest.mock('../Editor/EditorSlice')
 
 
 describe("When logged in and user owns project", () => {
   let store;
   let saveButton;
+  const project = {
+    identifier: "hello-world-project",
+    components: [],
+    image_list: [],
+    user_id: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
+  }
+  const user = {
+    access_token: "39a09671-be55-4847-baf5-8919a0c24a25",
+    profile: {
+      user: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
+    }
+  }
 
   beforeEach(() => {
     const middlewares = []
     const mockStore = configureStore(middlewares)
     const initialState = {
       editor: {
-        project: {
-          identifier: "hello-world-project",
-          components: [],
-          image_list: [],
-          user_id: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
-        },
-        projectLoaded: true,
+        project: project,
+        projectLoaded: 'success',
       },
       auth: {
-        user: {
-          access_token: "39a09671-be55-4847-baf5-8919a0c24a25",
-          profile: {
-            user: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
-          }
-        }
+        user: user
       }
     }
     store = mockStore(initialState);
@@ -54,31 +54,12 @@ describe("When logged in and user owns project", () => {
       expect(saveButton).toBeInTheDocument();
   })
 
-  test("Clicking save button sends PUT request to correct endpoint", () => {
-    axios.put.mockImplementationOnce(() => Promise.resolve({}))
+  test("Clicking save dispatches saveProject with correct parameters", async () => {
+    const saveAction = {type: 'SAVE_PROJECT' }
+    saveProject.mockImplementationOnce(() => (saveAction))
     fireEvent.click(saveButton)
-    const api_host = process.env.REACT_APP_API_ENDPOINT;
-    const access_token = "39a09671-be55-4847-baf5-8919a0c24a25"
-    const user_id = "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
-    const project = {"components": [], "image_list": [], "identifier": "hello-world-project", "user_id": user_id}
-    const headers = {"headers": {"Accept": "application/json", "Authorization": access_token}}
-    expect(axios.put).toHaveBeenCalledWith(`${api_host}/api/projects/hello-world-project`, {"project": project}, headers)
-  })
-
-  test("Clicking save button dispatches project once returned", async () => {
-    const project = {"components": [], "identifier": "hello-world-project", "user_id": "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"}
-    axios.put.mockImplementationOnce(() => Promise.resolve({ status: 200, data: project }))
-    fireEvent.click(saveButton)
-    await new Promise(process.nextTick);
-    const actions = store.getActions();
-    const expectedPayload = { type: 'editor/setProject', payload: project }
-    expect(actions).toEqual([expectedPayload])
-  })
-
-  test("Successful save prompts success message", async () => {
-    axios.put.mockImplementationOnce(() => Promise.resolve({ status: 200, data: {}}))
-    fireEvent.click(saveButton)
-    await waitFor(() => expect(showSavedMessage).toHaveBeenCalled())
+    await waitFor(() => expect(saveProject).toHaveBeenCalledWith({project, user, autosave: false}))
+    expect(store.getActions()[0]).toEqual(saveAction)
   })
 
   test("Renders project gallery link", () => {
@@ -97,26 +78,29 @@ describe("When logged in and user owns project", () => {
 describe("When logged in and no project identifier", () => {
   let store;
   let saveButton;
+  const project = {
+    components: [],
+    image_list: [],
+    user_id: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
+  }
+
+  const user = {
+    access_token: "39a09671-be55-4847-baf5-8919a0c24a25",
+    profile: {
+      user: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
+    }
+  }
 
   beforeEach(() => {
     const middlewares = []
     const mockStore = configureStore(middlewares)
     const initialState = {
       editor: {
-        project: {
-          components: [],
-          image_list: [],
-          user_id: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
-        },
-        projectLoaded: true,
+        project: project,
+        projectLoaded: 'success',
       },
       auth: {
-        user: {
-          access_token: "39a09671-be55-4847-baf5-8919a0c24a25",
-          profile: {
-            user: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
-          }
-        }
+        user: user
       }
     }
     store = mockStore(initialState);
@@ -132,27 +116,9 @@ describe("When logged in and no project identifier", () => {
     expect(saveButton).toBeInTheDocument()
   })
 
-  test("Clicking save creates new project", () => {
-    const project = {"components": [], "identifier": "hello-world-project", "user_id": "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"}
-    axios.post.mockImplementationOnce(() => Promise.resolve({ status: 200, data: project}))
-    fireEvent.click(saveButton)
-    const api_host = process.env.REACT_APP_API_ENDPOINT;
-    const access_token = "39a09671-be55-4847-baf5-8919a0c24a25"
-    const user_id = "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
-    const new_project = {"components": [], "image_list": [], "user_id": user_id}
-    const headers = {"headers": {"Accept": "application/json", "Authorization": access_token}}
-    expect(axios.post).toHaveBeenCalledWith(`${api_host}/api/projects`, {"project": new_project}, headers)
-  })
-
   test('Project name is shown', () => {
     expect(screen.queryByRole('textbox')).toBeInTheDocument()
    })
-   
-  test("Successful save prompts success message", async () => {
-    axios.post.mockImplementationOnce(() => Promise.resolve({ status: 200, data: {}}))
-    fireEvent.click(saveButton)
-    await waitFor(() => expect(showSavedMessage).toHaveBeenCalled())
-  })
 })
 
 describe("When logged in and user does not own project", () => {
@@ -164,12 +130,13 @@ describe("When logged in and user does not own project", () => {
     const initialState = {
       editor: {
         project: {
+          name: 'My first project',
           identifier: "hello-world-project",
           components: [],
           image_list: [],
           user_id: "b48e70e2-d9ed-4a59-aee5-fc7cf09dbfaf"
         },
-        projectLoaded: true,
+        projectLoaded: 'success',
       },
       auth: {
         user: {
@@ -197,7 +164,7 @@ describe("When logged in and user does not own project", () => {
   })
 
   test('Project name is shown', () => {
-    expect(screen.queryByText('header.newProject')).toBeInTheDocument()
+    expect(screen.queryByText('My first project')).toBeInTheDocument()
   })
 })
 
@@ -213,7 +180,7 @@ describe("When not logged in", () => {
             components: [],
             image_list: [],
           },
-          projectLoaded: true,
+          projectLoaded: 'success',
         },
         auth: {
           user: null
@@ -248,7 +215,7 @@ describe('When no project loaded', () => {
     const initialState = {
         editor: {
           project: {},
-          projectLoaded: false,
+          projectLoaded: 'idle',
         },
         auth: {
           user: "5254370e-26d2-4c8a-9526-8dbafea43aa9"