+++ src/components/Login/Login.js
@@ -14,7 +14,7 @@ const Login = (props) => {
   const onLoginButtonClick = (event) => {
     event.preventDefault();
     localStorage.setItem('location', location.pathname)
-    localStorage.setItem('project', JSON.stringify(project))
+    localStorage.setItem(project.identifier || 'project', JSON.stringify(project))
     userManager.signinRedirect();
   }