+++ src/utils/apiCallHandler.js
@@ -24,8 +24,8 @@ const headers = (accessToken) => {
     return {headers: headersHash}
 }
 
-export const createProject = async (accessToken) => {
-  return await post(`${host}/api/projects`, {}, headers(accessToken));
+export const createProject = async (project, accessToken) => {
+  return await post(`${host}/api/projects`, {project: project}, headers(accessToken));
 }
 
 export const deleteProject = async (identifier, accessToken) => {
@@ -50,10 +50,6 @@ export const readProject = async (projectIdentifier) => {
 
 export const readProjectList = async (accessToken) => {
   return await get(`${host}/api/projects`, headers(accessToken));
-}
-
-export const saveProject = async (project, accessToken) => {
-  return await post(`${host}/api/projects`, {project: project}, headers(accessToken));
 }
 
 export const updateProject = async (project, accessToken) => {